feat(products): prevent deletion of categories with associated products

+++ src/controllers/category.controller.js
@@ -1,5 +1,5 @@
 import { StatusCodes } from "http-status-codes";
-import { Category } from "../models/index.js";
+import { Category, Product } from "../models/index.js";
 import { sendResponse } from "../utils/index.js";
 import { BadRequestError, NotFoundError } from "../errors/index.js";
 import slugify from "slugify";
@@ -94,15 +94,15 @@ export const deleteCategory = async (req, res, next) => {
 			throw new NotFoundError(`No category found with slug: ${slug}`);
 		}
 
-		// Check for products in the associated category
-		// const productsInCategory = await Product.countDocuments({
-		// 	category: category._id,
-		// });
-		// if (productsInCategory > 0) {
-		// 	throw new BadRequestError(
-		// 		"Cannot delete category because it has associated products. Please re-assign or delete those products first."
-		// 	);
-		// }
+		const productsExists = await Product.countDocuments({
+			category: category._id,
+		});
+		if (productsExists > 0) {
+			throw new BadRequestError(
+				"Cannot delete category because it has associated products. Please re-assign or delete those products first."
+			);
+		}
+
 		await Category.findByIdAndDelete(category._id);
 
 		return res.status(StatusCodes.OK).json(
